+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - "Python: Division" HackerRank challenge
 - "Python: Loops" HackerRank challenge
 - "Python: Write a Function" HackerRank challenge
+- "Python: Print" HackerRank challenge
 
 ### Added
 - Change description